+++ tests/Integration/Internal/Module/Migration/Fixtures/myTestModuleWithoutMigrations/metadata.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+/**
+ * Metadata version
+ */
+$sMetadataVersion = '2.1';
+
+/**
+ * Module information
+ */
+$aModule = array(
+    'id'          => 'myTestModuleWithoutMigrations',
+    'title'       => 'myTestModuleWithoutMigrations',
+    'description' => 'myTestModuleWithoutMigrations',
+    'thumbnail'   => 'picture.png',
+    'version'     => '1.0',
+    'author'      => 'OXID eSales AG'
+);